Backported fix for CORE-3067: Objects not unmapped when shared memory is closed
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/branches/B2_5_Release@51319 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/isc_sync.cpp
@@ -2312,7 +2312,7 @@ void ISC_unmap_object(ISC_STATUS* status_vector,
 		return;
 	}
 #endif
-	const ULONG page_size = (ULONG) ps;
+	const size_t page_size = (ULONG) ps;
 
 /* Compute the start and end page-aligned addresses which
    contain the mapped object. */
